+++ app/src/organisms/ProtocolSetup/RunSetupCard/ProceedToRunCta.tsx
@@ -1,5 +1,6 @@
 import * as React from 'react'
 import { NavLink } from 'react-router-dom'
+import { useModuleMatchResults, useProtocolCalibrationStatus } from './hooks'
 import { useTranslation } from 'react-i18next'
 import {
   Flex,
@@ -9,17 +10,12 @@ import {
   JUSTIFY_CENTER,
 } from '@opentrons/components'
 import { useTrackEvent } from '../../../redux/analytics'
-import { ConfirmAttachmentModal } from '../../Devices/ModuleCard/ConfirmAttachmentModal'
-import { useHeaterShakerFromProtocol } from '../../Devices/ModuleCard/hooks'
-import { useModuleMatchResults, useProtocolCalibrationStatus } from './hooks'
 
 export const ProceedToRunCta = (): JSX.Element | null => {
   const { t } = useTranslation('protocol_setup')
   const [targetProps, tooltipProps] = useHoverTooltip()
   const moduleMatchResults = useModuleMatchResults()
   const trackEvent = useTrackEvent()
-  const heaterShaker = useHeaterShakerFromProtocol()
-  const isHeaterShakerInProtocol = heaterShaker != null
   const isEverythingCalibrated = useProtocolCalibrationStatus().complete
   const { missingModuleIds } = moduleMatchResults
   const calibrationIncomplete =
@@ -28,11 +24,6 @@ export const ProceedToRunCta = (): JSX.Element | null => {
     missingModuleIds.length > 0 && isEverythingCalibrated
   const moduleAndCalibrationIncomplete =
     missingModuleIds.length > 0 && !isEverythingCalibrated
-  const [
-    showConfirmAttachModal,
-    setShowConfirmAttachmentModal,
-  ] = React.useState<boolean>(false)
-
   let proceedToRunDisabledReason = null
   if (moduleAndCalibrationIncomplete) {
     proceedToRunDisabledReason = t(
@@ -43,32 +34,19 @@ export const ProceedToRunCta = (): JSX.Element | null => {
   } else if (moduleSetupIncomplete) {
     proceedToRunDisabledReason = t('run_disabled_modules_not_connected')
   }
-  const buttonInsteadOfNavLink =
-    isHeaterShakerInProtocol || proceedToRunDisabledReason != null
 
-  const LinkComponent = buttonInsteadOfNavLink ? 'button' : NavLink
-  const linkProps = buttonInsteadOfNavLink ? {} : { to: '/run' }
+  const LinkComponent = proceedToRunDisabledReason != null ? 'button' : NavLink
+  const linkProps = proceedToRunDisabledReason != null ? {} : { to: '/run' }
 
   return (
     <Flex justifyContent={JUSTIFY_CENTER}>
-      {showConfirmAttachModal && (
-        <ConfirmAttachmentModal
-          onCloseClick={() => setShowConfirmAttachmentModal(false)}
-          isProceedToRunModal={true}
-          shakerValue={null}
-        />
-      )}
       <NewPrimaryBtn
         role="button"
         title={t('proceed_to_run')}
         disabled={proceedToRunDisabledReason != null}
         as={LinkComponent}
         id={'LabwareSetup_proceedToRunButton'}
-        onClick={
-          isHeaterShakerInProtocol
-            ? () => setShowConfirmAttachmentModal(true)
-            : () => trackEvent({ name: 'proceedToRun', properties: {} })
-        }
+        onClick={() => trackEvent({ name: 'proceedToRun', properties: {} })}
         {...linkProps}
         {...targetProps}
       >
